tripbot/wsgi.py
"""
WSGI config for tripbot project.

It exposes the WSGI callable as a module-level variable named ``application``.

For more information on this file, see
https://docs.djangoproject.com/en/5.2/howto/deployment/wsgi/
"""
import dotenv
import os

from django.core.wsgi import get_wsgi_application

os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'tripbot.settings')
dotenv.load_dotenv()
application = get_wsgi_application()
