requirements.txt
line-bot-sdk>=3.11.0

Flask==3.0.3
gunicorn==23.0.0
Werkzeug==3.0.3

ag2==0.9.5
aiohappyeyeballs==2.6.1
aiohttp
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.0
asyncer==0.0.8
attrs==25.3.0
autogen==0.9.5
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
dataclasses-json==0.6.7
diskcache==5.6.3
distro==1.9.0
Django==4.2.11
django-cors-headers==4.7.0
djangorestframework==3.16.0
docker==7.1.0
eval_type_backport==0.2.2
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.5.1
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.33.2
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kagglehub==0.3.12
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.69
langchain-openai==0.3.28
langchain-text-splitters==0.3.8
langsmith==0.4.8
litellm
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
multidict==6.6.3
mypy_extensions==1.1.0
numpy==2.3.1
openai==1.93.0
orjson==3.11.0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
propcache==0.3.2
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
# pywin32==310
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.26.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.1.0
tiktoken==0.9.0
together==1.5.21
tokenizers==0.21.2
tqdm==4.67.1
typer==0.15.4
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
whitenoise>=6.0
